Rename all instances of client to human.
